[#440] iRODS 3.0 Release activities

+++ jargon-core/src/main/java/org/irods/jargon/core/pub/DataObjectAOImpl.java
@@ -814,7 +814,7 @@ public final class DataObjectAOImpl extends FileCatalogObjectAOImpl implements
 			return;
 		}
 
-		final long lengthFromIrodsResponse = temp.getIntValue();
+		final long lengthFromIrodsResponse = temp.getLongValue();
 
 		log.info("transfer length is:", lengthFromIrodsResponse);
